backend/requirements.txt
flask
requests
python-dotenv
PyMuPDF
flask-cors